+++ jest.config.mjs
@@ -13,4 +13,5 @@ export default {
       },
     ],
   },
+  testMatch: ["**/tests/**/*.test.ts"],
 };